pull out search box into new component
