+++ app/(public)/music/index.tsx
@@ -1,14 +1,12 @@
-import { Image, Pressable, ScrollView, Text, View, Alert } from "react-native";
+import { Image, Pressable, ScrollView, Text, View } from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
 import { router } from "expo-router";
 import { useEffect, useState } from "react";
 import { getMusic } from "../../../utils/api";
 import { Music } from "../../../types/front-end";
-import { TextInput } from "react-native-gesture-handler";
-import { Ionicons } from "@expo/vector-icons";
+import SearchDropDown from "../../../components/SearchDropDown";
 
 const Albums = () => {
-  const [searchText, setSearchText] = useState("");
   const [music, setMusic] = useState<Music[]>([]);
   useEffect(() => {
     const doThis = async () => {
@@ -17,16 +15,6 @@ const Albums = () => {
     };
     doThis();
   }, []);
-
-  const handleSearchSubmit = () => {
-    if (searchText) {
-      console.log(`Search for ${searchText}`);
-    } else {
-      Alert.alert("Incomplete Search", "Please add in text before searching", [
-        { text: "OK", onPress: () => console.log("close") },
-      ]);
-    }
-  };
 
   return (
     <SafeAreaView>
@@ -37,22 +25,7 @@ const Albums = () => {
           resizeMode="center"
         />
       </View>
-      <View className="flex flex-row items-center justify-items-center">
-        <TextInput
-          className="border-2 focus:border-green-600 m-5 p-3 rounded w-[75%]"
-          placeholder="Search"
-          value={searchText}
-          onChangeText={setSearchText}
-        />
-        <Pressable
-          className="bg-black border-2 border-black p-3 rounded"
-          onPress={handleSearchSubmit}
-        >
-          <Text>
-            <Ionicons name="search-outline" size={24} color="white" />
-          </Text>
-        </Pressable>
-      </View>
+      <SearchDropDown />
 
       <ScrollView>
         <View className="flex flex-row flex-wrap justify-between bg-white">
